# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateOAuthClientRequest:
    properties:
      access_token_ttl:
        example: 900
        type: integer
      allowed_grant_types:
        example:
        - authorization_code
        - refresh_token
        items:
          type: string
        minItems: 1
        type: array
      allowed_scopes:
        example:
        - openid
        - profile
        - email
        items:
          type: string
        minItems: 1
        type: array
      client_type:
        enum:
        - confidential
        - public
        example: confidential
        type: string
      default_scopes:
        example:
        - openid
        - profile
        items:
          type: string
        type: array
      description:
        example: My OAuth client application
        type: string
      first_party:
        example: false
        type: boolean
      id_token_ttl:
        example: 3600
        type: integer
      logo_url:
        example: https://example.com/logo.png
        type: string
      name:
        example: My Application
        maxLength: 100
        minLength: 3
        type: string
      redirect_uris:
        example:
        - https://example.com/callback
        items:
          type: string
        type: array
      refresh_token_ttl:
        example: 604800
        type: integer
      require_consent:
        example: true
        type: boolean
      require_pkce:
        example: true
        type: boolean
    required:
    - allowed_grant_types
    - allowed_scopes
    - client_type
    - name
    type: object
  models.CreateOAuthClientResponse:
    properties:
      client:
        $ref: '#/definitions/models.OAuthClient'
      client_secret:
        example: client_secret_abc123xyz789
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      details:
        description: Additional error details
        example: Email field is required
        type: string
      error:
        description: HTTP error status text
        example: Bad Request
        type: string
      message:
        description: Human-readable error message
        example: Invalid request parameters
        type: string
    type: object
...
  models.OAuthClient:
    properties:
      access_token_ttl:
        example: 900
        type: integer
      allowed_grant_types:
        example:
        - authorization_code
        - refresh_token
        items:
          type: string
        type: array
      allowed_scopes:
        example:
        - openid
        - profile
        - email
        items:
          type: string
        type: array
      client_id:
        example: my_client_app_123
        type: string
      client_type:
        example: confidential
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      default_scopes:
        example:
        - openid
        - profile
        items:
          type: string
        type: array
      description:
        example: My OAuth client application
        type: string
      first_party:
        example: false
        type: boolean
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      id_token_ttl:
        example: 3600
        type: integer
      is_active:
        example: true
        type: boolean
      logo_url:
        example: https://example.com/logo.png
        type: string
      name:
        example: My Application
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      redirect_uris:
        example:
        - https://example.com/callback
        items:
          type: string
        type: array
      refresh_token_ttl:
        example: 604800
        type: integer
      require_consent:
        example: true
        type: boolean
      require_pkce:
        example: true
        type: boolean
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
  models.UpdateOAuthClientRequest:
    properties:
      access_token_ttl:
        example: 900
        type: integer
      allowed_grant_types:
        example:
        - authorization_code
        - refresh_token
        items:
          type: string
        minItems: 1
        type: array
      allowed_scopes:
        example:
        - openid
        - profile
        - email
        items:
          type: string
        minItems: 1
        type: array
      default_scopes:
        example:
        - openid
        - profile
        items:
          type: string
        type: array
      description:
        example: Updated description
        type: string
      id_token_ttl:
        example: 3600
        type: integer
      is_active:
        example: true
        type: boolean
      logo_url:
        example: https://example.com/new-logo.png
        type: string
      name:
        example: My Updated Application
        maxLength: 100
        minLength: 3
        type: string
      redirect_uris:
        example:
        - https://example.com/callback
        items:
          type: string
        type: array
      refresh_token_ttl:
        example: 604800
        type: integer
      require_consent:
        example: true
        type: boolean
      require_pkce:
        example: true
        type: boolean
    type: object
...
host: localhost:8811
info:
  contact:
    email: maksbalashov@gmail.com
    name: API Support
  description: |-
    Centralized authentication and authorization system for microservices ecosystem.

    ## Overview
    Auth Gateway provides a complete authentication and authorization solution including:
    - User registration and authentication (email/password, OAuth, passwordless)
    - Two-factor authentication (TOTP)
    - API key management for service-to-service communication
    - Role-based access control (RBAC)
    - Session management
    - Webhooks for event notifications

    ## Authentication
    Most endpoints require authentication via JWT Bearer token or API key.

    ### JWT Authentication
    Include the access token in the Authorization header: `Authorization: Bearer {token}`

    ### API Key Authentication
    Include the API key in the X-API-Key header: `X-API-Key: agw_{key}`
  title: Auth Gateway API
  version: "1.0"
paths:
...
  /api/admin/oauth/clients:
    get:
      description: Get paginated list of OAuth 2.0 clients (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by owner ID
        in: query
        name: owner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response with clients, total, page, per_page
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List OAuth clients
      tags:
      - Admin - OAuth Clients
    post:
      consumes:
      - application/json
      description: Create a new OAuth 2.0 client application (admin only)
      parameters:
      - description: Client creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateOAuthClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateOAuthClientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create OAuth client
      tags:
      - Admin - OAuth Clients
  /api/admin/oauth/clients/{id}:
    delete:
      description: Soft-delete an OAuth 2.0 client (admin only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete OAuth client
      tags:
      - Admin - OAuth Clients
    get:
      description: Get detailed information about a specific OAuth client (admin only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OAuthClient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get OAuth client
      tags:
      - Admin - OAuth Clients
    put:
      consumes:
      - application/json
      description: Update OAuth 2.0 client information (admin only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Client update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateOAuthClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OAuthClient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update OAuth client
      tags:
      - Admin - OAuth Clients
  /api/admin/oauth/clients/{id}/consents:
    get:
      description: Get all user consents for a specific OAuth client (admin only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response with consents array
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List client consents
      tags:
      - Admin - OAuth Consents
  /api/admin/oauth/clients/{id}/consents/{user_id}:
    delete:
      description: Revoke a user's consent for a specific OAuth client (admin action)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke user consent
      tags:
      - Admin - OAuth Consents
  /api/admin/oauth/clients/{id}/rotate-secret:
    post:
      description: Generate a new client secret for an OAuth 2.0 client (admin only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response with client_secret
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rotate client secret
      tags:
      - Admin - OAuth Clients
  /api/admin/oauth/scopes:
    get:
      description: Get list of all OAuth 2.0 scopes (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Response with scopes array
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List OAuth scopes
      tags:
      - Admin - OAuth Scopes
    post:
      consumes:
      - application/json
      description: Create a custom OAuth 2.0 scope (admin only, is_system=false)
      parameters:
      - description: Scope data with name, display_name, description
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OAuthScope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create OAuth scope
      tags:
      - Admin - OAuth Scopes
  /api/admin/oauth/scopes/{id}:
    delete:
      description: Delete a custom OAuth 2.0 scope (admin only, cannot delete system
        scopes)
      parameters:
      - description: Scope ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete OAuth scope
      tags:
      - Admin - OAuth Scopes
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key for service-to-service authentication. Format: "agw_{key}"'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token for user authentication. Format: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"